#11 Added unit tests for Child, Tasks

+++ app/child/child-controller.js
@@ -2,8 +2,7 @@
 
 var Child = require('./child-model'),
   Score = require('../score/score-model'),
-  Payment = require('../score/payment-model'),
-  HttpError = require('../util/http-error');
+  Payment = require('../score/payment-model');
 
 /**
  * delete a child and its records
@@ -12,22 +11,36 @@ var Child = require('./child-model'),
  * @param  {Function} callback [description]
  */
 exports.deleteChild = function(userId, childId, callback) {
+  if (userId === null) {
+    return callback(new Error('Unauthorised.'));
+  }
   Child.findOneAndRemove({
     _user: userId,
     _id: childId
-  }, function(err) {
+  }, function(err, child) {
     if (err) {
       return callback(err);
+    }
+    if (!child) {
+      return callback(null, false);
     }
     Score.remove({
       _user: userId,
       _child: childId
-    }, function(err) {});
-    Payment.remove({
-      _user: userId,
-      _child: childId
-    }, function(err) {});
-    return callback(null);
+    }, function(err) {
+      if (err) {
+        return callback(err);
+      }
+      Payment.remove({
+        _user: userId,
+        _child: childId
+      }, function(err) {
+        if (err) {
+          return callback(err);
+        }
+        return callback(null, true);
+      });
+    });
   });
 };
 
@@ -39,6 +52,9 @@ exports.deleteChild = function(userId, childId, callback) {
  * @param  {Function} callback [description]
  */
 exports.updateTasks = function(userId, childId, tasks, callback) {
+  if (userId === null) {
+    return callback(new Error('Unauthorised.'));
+  }
   var uniqueTasks = tasks.filter(function(item, pos) {
     return tasks.indexOf(item) === pos;
   });
@@ -57,7 +73,7 @@ exports.updateTasks = function(userId, childId, tasks, callback) {
       return callback(err);
     }
     if (!child) {
-      return callback(new HttpError(404, 'Child not found.'));
+      return callback(null, false);
     }
     return callback(null, child.tasks);
   });
